move client to modules:client
