--- client/src/main/java/org/example/event/sourcing/order/poc/client/order/exception/BadRequestException.java
@@ -1,9 +1,0 @@
-package org.example.event.sourcing.order.poc.client.order.exception;
-
-public class BadRequestException extends RuntimeException {
-
-    public BadRequestException(String message) {
-        super(message);
-    }
-    
-}
