--- client/src/main/java/org/example/event/sourcing/order/poc/client/order/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package org.example.event.sourcing.order.poc.client.order.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
