--- client/src/main/java/org/example/event/sourcing/order/poc/client/order/model/ResourceName.java
@@ -1,7 +1,0 @@
-package org.example.event.sourcing.order.poc.client.order.model;
-
-public enum ResourceName {
-
-    ORDER, SHIPMENT, PAYMENT
-
-}
