--- client/src/main/java/org/example/event/sourcing/order/poc/client/order/model/V1OrderStatus.java
@@ -1,7 +1,0 @@
-package org.example.event.sourcing.order.poc.client.order.model;
-
-public enum V1OrderStatus {
-
-    CREATED, PREPARING, IN_LOGISTICS, PICKED_UP, FINISHED
-
-}
