--- client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-org.example.event.sourcing.order.poc.client.order.config.OrderQueryClientConfig
-org.example.event.sourcing.order.poc.client.order.config.OrderCommandClientConfig
