+++ modules/client/src/main/java/org/example/event/sourcing/order/poc/modules/client/order/exception/BadRequestException.java
@@ -1,0 +1,9 @@
+package org.example.event.sourcing.order.poc.modules.client.order.exception;
+
+public class BadRequestException extends RuntimeException {
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+    
+}
