+++ modules/client/src/main/java/org/example/event/sourcing/order/poc/modules/client/order/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package org.example.event.sourcing.order.poc.modules.client.order.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
