+++ modules/client/src/main/java/org/example/event/sourcing/order/poc/modules/client/order/model/ResourceName.java
@@ -1,0 +1,7 @@
+package org.example.event.sourcing.order.poc.modules.client.order.model;
+
+public enum ResourceName {
+
+    ORDER, SHIPMENT, PAYMENT
+
+}
