+++ modules/client/src/main/java/org/example/event/sourcing/order/poc/modules/client/order/model/V1OrderStatus.java
@@ -1,0 +1,7 @@
+package org.example.event.sourcing.order.poc.modules.client.order.model;
+
+public enum V1OrderStatus {
+
+    CREATED, PREPARING, IN_LOGISTICS, PICKED_UP, FINISHED
+
+}
