+++ order/command-side/build.gradle
@@ -3,7 +3,7 @@ jib.to.image = "noahhsu/espoc-order-${project.name}"
 dependencies {
     implementation project(":common")
     implementation project(":event")
-    implementation project(":client")
+    implementation project(":modules:client")
     implementation project(":observation")
     implementation project(":modules:idempotency")
