+++ order/command-side/src/main/java/org/example/event/sourcing/order/poc/command/order/service/OrderService.java
@@ -2,22 +2,19 @@ package org.example.event.sourcing.order.poc.command.order.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.client.order.OrderQueryClient;
-import org.example.event.sourcing.order.poc.client.order.exception.ResourceNotFoundException;
-import org.example.event.sourcing.order.poc.client.order.model.V1Order;
-import org.example.event.sourcing.order.poc.client.order.model.V1OrderStatus;
+import org.example.event.sourcing.order.poc.modules.client.order.OrderQueryClient;
+import org.example.event.sourcing.order.poc.modules.client.order.exception.ResourceNotFoundException;
+import org.example.event.sourcing.order.poc.modules.client.order.model.V1Order;
+import org.example.event.sourcing.order.poc.modules.client.order.model.V1OrderStatus;
 import org.example.event.sourcing.order.poc.command.order.producer.OrderEventProducer;
 import org.example.event.sourcing.order.poc.common.model.Order;
 import org.example.event.sourcing.order.poc.event.model.OrderEvent;
 import org.example.event.sourcing.order.poc.observation.annotation.LogInfo;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.ErrorResponseException;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import java.time.Instant;
 import java.util.Optional;
 import java.util.Random;
