+++ shipment/event-handler/build.gradle
@@ -2,7 +2,7 @@ jib.to.image = "noahhsu/espoc-shipment-${project.name}"
 
 dependencies {
     implementation project(":common")
-    implementation project(":client")
+    implementation project(":modules:client")
     implementation project(":event")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
